+++ pyzoo/zoo/zouwu/autots/forecast.py
@@ -14,7 +14,7 @@
 # limitations under the License.
 #
 
-from zoo.automl.regression.time_sequence_predictor import TimeSequencePredictor
+from zoo.zouwu.regression.time_sequence_predictor import TimeSequencePredictor
 from zoo.zouwu.config.recipe import *
 from zoo.zouwu.pipeline.time_sequence import load_ts_pipeline
